Use db.Log to log database errors, Fix slice resolvers
looping through a slice of resolvers now works for most things, I will
likely find other places where thats broken too.

Whenever an error occurs in a resolver, log it to the database.
Eventually, success messages will also be logged

+++ resolvers/root.go
@@ -32,7 +32,7 @@ func GetSearch(args arguments.Search) (name string, id int, err error) {
 	return
 }
 
-func GetLimitOffset(args arguments.Connection) (int, int, error) {
+func GetLimitOffset(args arguments.Connection) (uint64, uint64, error) {
 	offset := 0
 	limit := 20
 
@@ -49,5 +49,5 @@ func GetLimitOffset(args arguments.Connection) (int, int, error) {
 		limit = int(*args.First)
 	}
 
-	return limit, offset, nil
+	return uint64(limit), uint64(offset), nil
 }
